pyproject.toml
[tool.poetry]
name = "windchimes-backend"
version = "0.1.0"
description = ""
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
uvicorn = "^0.34.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.256.1"}
pydantic-settings = "^2.7.1"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
auth0-python = "^4.7.2"
apscheduler = "^3.11.0"
pytubefix = "^8.11.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
